Refactor table of contents to use one component and simplify useHeadings
